update installer.sh to use ifdata for iface info
somewhere along the line, someone changed the output format of ifconfig such that (at least on my ubuntu 16.10 and deb 9 containers) the current method of parsing the ifconfig output for the ip addr and netmask breaks and returns nothing.

ifdata is meant for use in scripts and shortens up those lines a lot, and will probably be more robust against differing distros.  The downside is that it requires yet more libs to be installed, but then again we also do an entire make of nginx and sniproxy.  Tested & working on ubuntu 16.10 and deb 9 containers.